+++ app/controllers/product_variants_controller.rb
@@ -11,7 +11,7 @@ class ProductVariantsController < ApplicationController
     @brand = @product.brand
 
     if user_signed_in?
-      @possession = current_user.possessions.find_by(product_id: @product.id, product_variant_id: @variant.id)
+      @possessions = current_user.possessions.where(product_id: @product.id, product_variant_id: @variant.id)
       @bookmark = current_user.bookmarks.find_by(product_id: @product.id, product_variant_id: @variant.id)
       @prev_owned = current_user.prev_owneds.find_by(product_id: @product.id, product_variant_id: @variant.id)
       @setups = current_user.setups.includes(:possessions)