+++ db/migrate/20240311234305_add_created_at_to_possessions.rb
@@ -1,0 +1,5 @@
+class AddCreatedAtToPossessions < ActiveRecord::Migration[7.1]
+  def change
+    add_column :possessions, :created_at, :datetime
+  end
+end