+++ db/migrate/20240312002323_allow_multiple_possessions_of_one_product.rb
@@ -1,0 +1,6 @@
+class AllowMultiplePossessionsOfOneProduct < ActiveRecord::Migration[7.1]
+  def change
+    remove_index :possessions, name: 'idx_on_product_id_user_id_product_variant_id_9e45f66b35'
+    add_index :possessions, [:product_id, :product_variant_id, :user_id]
+  end
+end